adjust hbm2ddl mode, change car id to bigint

+++ src/main/resources/Car.sql
@@ -1,5 +1,5 @@
 CREATE TABLE Car (
-    id int(5) NOT NULL AUTO_INCREMENT,
+    id bigint(5) NOT NULL AUTO_INCREMENT,
     model varchar(50) DEFAULT NULL,
     PRIMARY KEY(id)
     );
